refactor: use new wrapper structs in tests
